scripts/create_model_card.py
#!/usr/bin/env python3
"""
Model card generation script for Hugging Face Hub uploads.
"""
import argparse
import json
import os
from datetime import datetime
from pathlib import Path
from typing import Dict, Any


def load_model_config(model_path: str) -> Dict[str, Any]:
    """Load model configuration from various possible sources."""
    model_path = Path(model_path)

    # Try to load from config.json
    config_file = model_path / "config.json"
    if config_file.exists():
        with open(config_file, 'r') as f:
            return json.load(f)

    # Try to load from model_config.json
    config_file = model_path / "model_config.json"
    if config_file.exists():
        with open(config_file, 'r') as f:
            return json.load(f)

    # Return empty config if none found
    return {}


def generate_model_card(model_path: str, model_type: str, output_file: str) -> None:
    """Generate a comprehensive model card for the trained model."""

    model_config = load_model_config(model_path)

    # Model card templates based on model type
    card_templates = {
        "clip": {
            "task": "multimodal-embedding",
            "description": "Fine-tuned CLIP model for e-commerce product understanding",
            "base_model": "openai/clip-vit-base-patch32"
        },
        "fusion": {
            "task": "text-embedding",
            "description": "Advanced embedding fusion model combining CLIP and SentenceTransformer",
            "base_model": "sentence-transformers/all-MiniLM-L6-v2"
        },
        "full_pipeline": {
            "task": "feature-extraction",
            "description": "Complete RAG pipeline with multimodal embeddings",
            "base_model": "multiple"
        }
    }

    template = card_templates.get(model_type, card_templates["clip"])

    model_card = f"""---
license: mit
base_model: {template["base_model"]}
tags:
- ecommerce
- rag
- multimodal
- {model_type}
- clip
- embedding
datasets:
- custom-ecommerce-dataset
language:
- en
metrics:
- recall
- precision
- bleu
- rouge
library_name: transformers
pipeline_tag: {template["task"]}
---

# E-commerce RAG {model_type.upper()} Model

## Model Description

{template["description"]} specifically trained for e-commerce product search and recommendation tasks.

This model is part of the E-commerce RAG Pipeline project, designed to understand and retrieve relevant product information from multimodal queries combining text descriptions and product images.

## Model Details

- **Model Type**: {model_type}
- **Base Model**: {template["base_model"]}
- **Training Date**: {datetime.now().strftime("%Y-%m-%d")}
- **Framework**: PyTorch + Transformers
- **Language**: English
- **License**: MIT

## Training Data

The model was trained on a curated e-commerce dataset containing:
- Product titles and descriptions
- Product images
- Category information
- Price and specification data
- Customer reviews and ratings

## Usage

```python
from transformers import CLIPModel, CLIPProcessor
from sentence_transformers import SentenceTransformer

# Load the model
model = CLIPModel.from_pretrained("your-username/ecommerce-rag-{model_type}")
processor = CLIPProcessor.from_pretrained("your-username/ecommerce-rag-{model_type}")

# Use for inference
text = "wireless bluetooth headphones"
inputs = processor(text=text, return_tensors="pt")
text_features = model.get_text_features(**inputs)
```

## Training Procedure

### Training Hyperparameters

- **Training regime**: {model_config.get("training_regime", "Fine-tuning")}
- **Learning rate**: {model_config.get("learning_rate", "2e-5")}
- **Batch size**: {model_config.get("batch_size", "16")}
- **Epochs**: {model_config.get("epochs", "5")}
- **Optimizer**: {model_config.get("optimizer", "AdamW")}

### Training Losses

- **Contrastive Loss**: Used for learning text-image correspondences
- **Domain Adaptation Loss**: Used for e-commerce domain specialization
- **Category Clustering Loss**: Used for better category understanding

## Evaluation

The model was evaluated on held-out e-commerce data with the following metrics:

- **Recall@1**: {model_config.get("recall_at_1", "TBD")}
- **Recall@5**: {model_config.get("recall_at_5", "TBD")}
- **Recall@10**: {model_config.get("recall_at_10", "TBD")}
- **BLEU Score**: {model_config.get("bleu_score", "TBD")}
- **ROUGE Score**: {model_config.get("rouge_score", "TBD")}

## Intended Use

This model is intended for:
- E-commerce product search
- Product recommendation systems
- Multimodal product understanding
- RAG-based question answering for products

## Limitations

- Trained specifically on e-commerce data
- May not generalize well to other domains
- Performance may vary with product categories not well-represented in training data

## Model Architecture

{json.dumps(model_config.get("architecture", {}), indent=2)}

## Citation

If you use this model, please cite:

```bibtex
@misc{{ecommerce-rag-{model_type},
  title={{E-commerce RAG {model_type.upper()} Model}},
  author={{Your Team}},
  year={{2024}},
  url={{https://huggingface.co/your-username/ecommerce-rag-{model_type}}}
}}
```

## Contact

For questions or issues, please contact: your-email@example.com

---

*This model was automatically uploaded via GitHub Actions CI/CD pipeline.*
"""

    # Write the model card
    with open(output_file, 'w', encoding='utf-8') as f:
        f.write(model_card)

    print(f"Model card generated successfully: {output_file}")


def main():
    parser = argparse.ArgumentParser(description="Generate model card for Hugging Face Hub")
    parser.add_argument("--model-path", required=True, help="Path to the trained model")
    parser.add_argument("--model-type", required=True, choices=["clip", "fusion", "full_pipeline"],
                       help="Type of model")
    parser.add_argument("--output-file", required=True, help="Output file path for the model card")

    args = parser.parse_args()

    generate_model_card(args.model_path, args.model_type, args.output_file)


if __name__ == "__main__":
    main()